Move Notification Permissions from WelcomeEntrant to EntrantHome, to handle case where a new user is made at WelcomeEntrant
